+++ src/main/java/com/arklok/nutra/helpers/ControllersHelper.java
@@ -1,0 +1,14 @@
+package com.arklok.nutra.helpers;
+
+import javafx.scene.control.Alert;
+
+public class ControllersHelper {
+
+    public static void showAlert(Alert.AlertType type, String title, String message) {
+        Alert alert = new Alert(type);
+        alert.setTitle(title);
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        alert.showAndWait();
+    }
+}